refactor: convert <MenuBar> component to typescript
