+++ client/src/features/menu/__tests__/MenuBar.test.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { QueryClientProvider } from 'react-query';
 import { render, screen } from '@testing-library/react';
 
@@ -7,11 +6,13 @@ import MenuBar from '../MenuBar';
 
 const onOpenHandler = jest.fn();
 const onCloseHandler = jest.fn();
+const isOpen = false;
+const onUploadOpenHandler = jest.fn();
 
 const renderInMock = () => {
   render(
     <QueryClientProvider client={queryClientMock}>
-      <MenuBar onOpen={onOpenHandler} onClose={onCloseHandler} />
+      <MenuBar onOpen={onOpenHandler} onClose={onCloseHandler} isOpen={isOpen} onUploadOpen={onUploadOpenHandler} />
     </QueryClientProvider>
   );
 };
